fix: add throw error to create func

+++ src/fs/create.js
@@ -1,6 +1,12 @@
 import fs from 'fs';
 
 const create = async () => {
+    const filePath = './fresh.txt';
+
+    if (fs.existsSync(filePath)) {
+        throw new Error('FS operation failed');
+    }
+
     try {
         await fs.promises.writeFile('./fresh.txt', 'I am fresh and young')
     } catch (err) {
